[fix] 테스트코드 작성

+++ app/src/main/java/com/tosslab/jandi/app/services/socket/to/SocketAnnouncementCreatedEvent.java
@@ -106,14 +106,14 @@ public class SocketAnnouncementCreatedEvent implements EventHistoryInfo {
     @JsonIgnoreProperties(ignoreUnknown = true)
     @JsonSerialize(include = JsonSerialize.Inclusion.NON_NULL)
     public static class Data {
-        private int topicId;
+        private long topicId;
         private Topic.Announcement announcement;
 
-        public int getTopicId() {
+        public long getTopicId() {
             return topicId;
         }
 
-        public void setTopicId(int topicId) {
+        public void setTopicId(long topicId) {
             this.topicId = topicId;
         }
